further breakdown nlp
