+++ src/dateutil.js
@@ -174,31 +174,36 @@ const dateutil = {
   }
 }
 
-dateutil.Time = function (hour, minute, second, millisecond) {
-  this.hour = hour
-  this.minute = minute
-  this.second = second
-  this.millisecond = millisecond || 0
-}
+class Time {
+  constructor (hour, minute, second, millisecond) {
+    this.hour = hour
+    this.minute = minute
+    this.second = second
+    this.millisecond = millisecond || 0
+  }
 
-dateutil.Time.prototype = {
-  constructor: dateutil.Time,
-  getHours: function () {
+  getHours () {
     return this.hour
-  },
-  getMinutes: function () {
+  }
+
+  getMinutes () {
     return this.minute
-  },
-  getSeconds: function () {
+  }
+
+  getSeconds () {
     return this.second
-  },
-  getMilliseconds: function () {
+  }
+
+  getMilliseconds () {
     return this.millisecond
-  },
-  getTime: function () {
+  }
+
+  getTime () {
     return ((this.hour * 60 * 60) + (this.minute * 60) + this.second) * 1000 +
       this.millisecond
   }
 }
+
+dateutil.Time = Time
 
 export default dateutil
