+++ src/totext.js
@@ -1,0 +1,407 @@
+import RRule from './rrule'
+import ENGLISH from './i18n'
+
+// =============================================================================
+// Helper functions
+// =============================================================================
+
+/**
+ * Return true if a value is in an array
+ */
+const contains = function (arr, val) {
+  return arr.indexOf(val) !== -1
+}
+
+// =============================================================================
+// ToText
+// =============================================================================
+
+/**
+ *
+ * @param {RRule} rrule
+ * Optional:
+ * @param {Function} gettext function
+ * @param {Object} language definition
+ * @constructor
+ */
+export default class ToText {
+  constructor (rrule, gettext, language) {
+    this.text = ''
+    this.language = language || ENGLISH
+    this.gettext = gettext || function (id) {
+      return id
+    }
+
+    this.rrule = rrule
+    this.freq = rrule.options.freq
+    this.options = rrule.options
+    this.origOptions = rrule.origOptions
+
+    if (this.origOptions.bymonthday) {
+      const bymonthday = [].concat(this.options.bymonthday)
+      const bynmonthday = [].concat(this.options.bynmonthday)
+
+      bymonthday.sort()
+      bynmonthday.sort()
+      bynmonthday.reverse()
+      // 1, 2, 3, .., -5, -4, -3, ..
+      this.bymonthday = bymonthday.concat(bynmonthday)
+      if (!this.bymonthday.length) this.bymonthday = null
+    }
+
+    if (this.origOptions.byweekday) {
+      const byweekday = !(this.origOptions.byweekday instanceof Array)
+        ? [this.origOptions.byweekday] : this.origOptions.byweekday
+      const days = String(byweekday)
+
+      this.byweekday = {
+        allWeeks: byweekday.filter(function (weekday) {
+          return !weekday.n
+        }),
+        someWeeks: byweekday.filter(function (weekday) {
+          return Boolean(weekday.n)
+        }),
+        isWeekdays: (
+          days.indexOf('MO') !== -1 &&
+          days.indexOf('TU') !== -1 &&
+          days.indexOf('WE') !== -1 &&
+          days.indexOf('TH') !== -1 &&
+          days.indexOf('FR') !== -1 &&
+          days.indexOf('SA') === -1 &&
+          days.indexOf('SU') === -1
+        )
+      }
+
+      const sortWeekDays = function (a, b) {
+        return a.weekday - b.weekday
+      }
+
+      this.byweekday.allWeeks.sort(sortWeekDays)
+      this.byweekday.someWeeks.sort(sortWeekDays)
+
+      if (!this.byweekday.allWeeks.length) this.byweekday.allWeeks = null
+      if (!this.byweekday.someWeeks.length) this.byweekday.someWeeks = null
+    } else {
+      this.byweekday = null
+    }
+  }
+
+  /**
+   * Test whether the rrule can be fully converted to text.
+   * @param {RRule} rrule
+   * @return {Boolean}
+   */
+  static isFullyConvertible (rrule) {
+    let canConvert = true
+
+    if (!(rrule.options.freq in ToText.IMPLEMENTED)) return false
+    if (rrule.origOptions.until && rrule.origOptions.count) return false
+
+    for (let key in rrule.origOptions) {
+      if (contains(['dtstart', 'wkst', 'freq'], key)) return true
+      if (!contains(ToText.IMPLEMENTED[rrule.options.freq], key)) return false
+    }
+
+    return canConvert
+  }
+
+  isFullyConvertible () {
+    return ToText.isFullyConvertible(this.rrule)
+  }
+
+  /**
+   * Perform the conversion. Only some of the frequencies are supported.
+   * If some of the rrule's options aren't supported, they'll
+   * be omitted from the output an "(~ approximate)" will be appended.
+   * @return {*}
+   */
+  toString () {
+    const gettext = this.gettext
+
+    if (!(this.options.freq in ToText.IMPLEMENTED)) {
+      return gettext('RRule error: Unable to fully convert this rrule to text')
+    }
+
+    this.text = [gettext('every')]
+    this[RRule.FREQUENCIES[this.options.freq]]()
+
+    if (this.options.until) {
+      this.add(gettext('until'))
+      const until = this.options.until
+      this.add(this.language.monthNames[until.getMonth()])
+        .add(until.getDate() + ',')
+        .add(until.getFullYear())
+    } else if (this.options.count) {
+      this.add(gettext('for'))
+        .add(this.options.count)
+        .add(this.plural(this.options.count)
+          ? gettext('times') : gettext('time'))
+    }
+
+    if (!this.isFullyConvertible()) this.add(gettext('(~ approximate)'))
+
+    return this.text.join('')
+  }
+
+  HOURLY () {
+    const gettext = this.gettext
+
+    if (this.options.interval !== 1) this.add(this.options.interval)
+
+    this.add(this.plural(this.options.interval)
+      ? gettext('hours') : gettext('hour'))
+  }
+
+  MINUTELY () {
+    const gettext = this.gettext
+
+    if (this.options.interval !== 1) this.add(this.options.interval)
+
+    this.add(this.plural(this.options.interval)
+      ? gettext('minutes') : gettext('minutes'))
+  }
+
+  DAILY () {
+    const gettext = this.gettext
+
+    if (this.options.interval !== 1) this.add(this.options.interval)
+
+    if (this.byweekday && this.byweekday.isWeekdays) {
+      this.add(this.plural(this.options.interval)
+        ? gettext('weekdays') : gettext('weekday'))
+    } else {
+      this.add(this.plural(this.options.interval)
+        ? gettext('days') : gettext('day'))
+    }
+
+    if (this.origOptions.bymonth) {
+      this.add(gettext('in'))
+      this._bymonth()
+    }
+
+    if (this.bymonthday) {
+      this._bymonthday()
+    } else if (this.byweekday) {
+      this._byweekday()
+    } else if (this.origOptions.byhour) {
+      this._byhour()
+    }
+  }
+
+  WEEKLY () {
+    const gettext = this.gettext
+
+    if (this.options.interval !== 1) {
+      this.add(this.options.interval)
+        .add(this.plural(this.options.interval)
+          ? gettext('weeks') : gettext('week'))
+    }
+
+    if (this.byweekday && this.byweekday.isWeekdays) {
+      if (this.options.interval === 1) {
+        this.add(this.plural(this.options.interval)
+          ? gettext('weekdays') : gettext('weekday'))
+      } else {
+        this.add(gettext('on')).add(gettext('weekdays'))
+      }
+    } else {
+      if (this.options.interval === 1) this.add(gettext('week'))
+
+      if (this.origOptions.bymonth) {
+        this.add(gettext('in'))
+        this._bymonth()
+      }
+
+      if (this.bymonthday) {
+        this._bymonthday()
+      } else if (this.byweekday) {
+        this._byweekday()
+      }
+    }
+  }
+
+  MONTHLY () {
+    const gettext = this.gettext
+
+    if (this.origOptions.bymonth) {
+      if (this.options.interval !== 1) {
+        this.add(this.options.interval).add(gettext('months'))
+        if (this.plural(this.options.interval)) this.add(gettext('in'))
+      } else {
+        // this.add(gettext('MONTH'))
+      }
+      this._bymonth()
+    } else {
+      if (this.options.interval !== 1) this.add(this.options.interval)
+      this.add(this.plural(this.options.interval)
+        ? gettext('months') : gettext('month'))
+    }
+    if (this.bymonthday) {
+      this._bymonthday()
+    } else if (this.byweekday && this.byweekday.isWeekdays) {
+      this.add(gettext('on')).add(gettext('weekdays'))
+    } else if (this.byweekday) {
+      this._byweekday()
+    }
+  }
+
+  YEARLY () {
+    const gettext = this.gettext
+
+    if (this.origOptions.bymonth) {
+      if (this.options.interval !== 1) {
+        this.add(this.options.interval)
+        this.add(gettext('years'))
+      } else {
+        // this.add(gettext('YEAR'))
+      }
+      this._bymonth()
+    } else {
+      if (this.options.interval !== 1) this.add(this.options.interval)
+      this.add(this.plural(this.options.interval)
+        ? gettext('years') : gettext('year'))
+    }
+
+    if (this.bymonthday) {
+      this._bymonthday()
+    } else if (this.byweekday) {
+      this._byweekday()
+    }
+
+    if (this.options.byyearday) {
+      this.add(gettext('on the'))
+        .add(this.list(this.options.byyearday, this.nth, gettext('and')))
+        .add(gettext('day'))
+    }
+
+    if (this.options.byweekno) {
+      this.add(gettext('in'))
+        .add(this.plural(this.options.byweekno.length) ? gettext('weeks') : gettext('week'))
+        .add(this.list(this.options.byweekno, null, gettext('and')))
+    }
+  }
+
+  _bymonthday () {
+    const gettext = this.gettext
+    if (this.byweekday && this.byweekday.allWeeks) {
+      this.add(gettext('on'))
+        .add(this.list(this.byweekday.allWeeks, this.weekdaytext, gettext('or')))
+        .add(gettext('the'))
+        .add(this.list(this.bymonthday, this.nth, gettext('or')))
+    } else {
+      this.add(gettext('on the'))
+        .add(this.list(this.bymonthday, this.nth, gettext('and')))
+    }
+    // this.add(gettext('DAY'))
+  }
+
+  _byweekday () {
+    const gettext = this.gettext
+    if (this.byweekday.allWeeks && !this.byweekday.isWeekdays) {
+      this.add(gettext('on'))
+        .add(this.list(this.byweekday.allWeeks, this.weekdaytext))
+    }
+
+    if (this.byweekday.someWeeks) {
+      if (this.byweekday.allWeeks) this.add(gettext('and'))
+
+      this.add(gettext('on the'))
+        .add(this.list(this.byweekday.someWeeks, this.weekdaytext, gettext('and')))
+    }
+  }
+
+  _byhour () {
+    const gettext = this.gettext
+
+    this.add(gettext('at'))
+      .add(this.list(this.origOptions.byhour, null, gettext('and')))
+  }
+
+  _bymonth () {
+    this.add(this.list(this.options.bymonth, this.monthtext, this.gettext('and')))
+  }
+
+  nth (n) {
+    let nth, npos
+    const gettext = this.gettext
+
+    if (n === -1) return gettext('last')
+
+    npos = Math.abs(n)
+    switch (npos) {
+      case 1:
+      case 21:
+      case 31:
+        nth = npos + gettext('st')
+        break
+      case 2:
+      case 22:
+        nth = npos + gettext('nd')
+        break
+      case 3:
+      case 23:
+        nth = npos + gettext('rd')
+        break
+      default:
+        nth = npos + gettext('th')
+    }
+
+    return n < 0 ? nth + ' ' + gettext('last') : nth
+  }
+
+  monthtext (m) {
+    return this.language.monthNames[m - 1]
+  }
+
+  weekdaytext (wday) {
+    const weekday = typeof wday === 'number' ? (wday + 1) % 7 : wday.getJsWeekday()
+    return (wday.n ? this.nth(wday.n) + ' ' : '') +
+      this.language.dayNames[weekday]
+  }
+
+  plural (n) {
+    return n % 100 !== 1
+  }
+
+  add (s) {
+    this.text.push(' ')
+    this.text.push(s)
+    return this
+  }
+
+  list (arr, callback, finalDelim, delim) {
+    if (!(arr instanceof Array)) {
+      arr = [arr]
+    }
+    const delimJoin = function (array, delimiter, finalDelimiter) {
+      let list = ''
+
+      for (let i = 0; i < array.length; i++) {
+        if (i !== 0) {
+          if (i === array.length - 1) {
+            list += ' ' + finalDelimiter + ' '
+          } else {
+            list += delimiter + ' '
+          }
+        }
+        list += array[i]
+      }
+      return list
+    }
+
+    delim = delim || ','
+    callback = callback || function (o) {
+      return o
+    }
+    const self = this
+    const realCallback = function (arg) {
+      return callback.call(self, arg)
+    }
+
+    if (finalDelim) {
+      return delimJoin(arr.map(realCallback), delim, finalDelim)
+    } else {
+      return arr.map(realCallback).join(delim + ' ')
+    }
+  }
+}
